requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.2
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
multidict==6.0.4
pycodestyle==2.10.0
pycparser==2.21
PyNaCl==1.5.0
requests==2.29.0
tomli==2.0.1
urllib3==1.26.15
yarl==1.9.2
